feat!: deprecate pre-tdec
