+++ packages/pre/src/characters/index.ts
@@ -1,4 +1,3 @@
 export * from './alice';
 export * from './bob';
 export * from './enrico';
-export * from './pre-recipient';
