+++ packages/pre/src/policy.ts
@@ -124,7 +124,7 @@ export class BlockchainPolicy {
   public async generatePreEnactedPolicy(
     ursulas: readonly Ursula[],
   ): Promise<PreEnactedPolicy> {
-    if (ursulas.length != this.verifiedKFrags.length) {
+    if (ursulas.length !== this.verifiedKFrags.length) {
       throw new Error(
         `Number of ursulas must match number of verified kFrags: ${this.verifiedKFrags.length}`,
       );
