+++ packages/pre/test/enrico.test.ts
@@ -1,22 +1,12 @@
 // Disabling because we want to access Alice.keyring which is a private property
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import {
-  bytesEqual,
-  fakeAlice,
-  fakeBob,
-  fromBytes,
-  reencryptKFrags,
-} from '@nucypher/test-utils';
+import { bytesEqual, fromBytes } from '@nucypher/test-utils';
 import { expect, test } from 'vitest';
 
-import {
-  ConditionExpression,
-  Enrico,
-  ERC721Ownership,
-  PolicyMessageKit,
-  RetrievalResult,
-  toBytes,
-} from '../src';
+import { Enrico, toBytes } from '../src';
+import { PolicyMessageKit, RetrievalResult } from '../src/kits';
+
+import { fakeAlice, fakeBob, reencryptKFrags } from './utils';
 
 test('enrico', () => {
   test('alice decrypts message encrypted by enrico', async () => {
@@ -26,7 +16,7 @@ test('enrico', () => {
 
     const policyKey = alice.getPolicyEncryptingKeyFromLabel(label);
     const enrico = new Enrico(policyKey);
-    const encrypted = enrico.encryptMessagePre(toBytes(message));
+    const encrypted = enrico.encryptMessage(toBytes(message));
 
     const aliceKeyring = (alice as any).keyring;
     const aliceSk = await aliceKeyring.getSecretKeyFromLabel(label);
@@ -44,7 +34,7 @@ test('enrico', () => {
 
     const plaintext = 'Plaintext message';
     const plaintextBytes = toBytes(plaintext);
-    const encrypted = enrico.encryptMessagePre(plaintextBytes);
+    const encrypted = enrico.encryptMessage(plaintextBytes);
 
     // Alice can decrypt capsule she created
     const aliceSk = await (alice as any).keyring.getSecretKeyFromLabel(label);
@@ -92,63 +82,5 @@ test('enrico', () => {
 
     const decrypted = bob.decrypt(pk);
     expect(bytesEqual(decrypted, plaintextBytes)).toBeTruthy();
-  });
-
-  test('enrico generates a message kit with conditions', async () => {
-    const label = 'fake-label';
-    const message = 'fake-message';
-    const alice = fakeAlice();
-
-    const policyKey = alice.getPolicyEncryptingKeyFromLabel(label);
-
-    const ownsBufficornNFT = new ERC721Ownership({
-      contractAddress: '0x1e988ba4692e52Bc50b375bcC8585b95c48AaD77',
-      parameters: [3591],
-      chain: 5,
-    });
-
-    const conditions = new ConditionExpression(ownsBufficornNFT);
-
-    const enrico = new Enrico(policyKey, undefined, conditions);
-    const encrypted = enrico.encryptMessagePre(toBytes(message));
-
-    const aliceKeyring = (alice as any).keyring;
-    const aliceSk = await aliceKeyring.getSecretKeyFromLabel(label);
-    const alicePlaintext = encrypted.decrypt(aliceSk);
-    expect(alicePlaintext).toEqual(alicePlaintext);
-  });
-
-  test('can overwrite conditions at encryption time', async () => {
-    const label = 'fake-label';
-    const message = 'fake-message';
-    const alice = fakeAlice();
-
-    const policyKey = alice.getPolicyEncryptingKeyFromLabel(label);
-
-    const ownsBufficornNFT = new ERC721Ownership({
-      contractAddress: '0x1e988ba4692e52Bc50b375bcC8585b95c48AaD77',
-      chain: 5,
-      parameters: [3591],
-    });
-
-    const ownsNonsenseNFT = new ERC721Ownership({
-      contractAddress: '0x1e988ba4692e52Bc50b375bcC8585b95c48AaD77',
-      chain: 5,
-      parameters: [6969],
-    });
-
-    const conditions = new ConditionExpression(ownsBufficornNFT);
-    const updatedConditions = new ConditionExpression(ownsNonsenseNFT);
-
-    const enrico = new Enrico(policyKey, undefined, conditions);
-    const encrypted = enrico.encryptMessagePre(
-      toBytes(message),
-      updatedConditions,
-    );
-
-    const aliceKeyring = (alice as any).keyring;
-    const aliceSk = await aliceKeyring.getSecretKeyFromLabel(label);
-    const alicePlaintext = encrypted.decrypt(aliceSk);
-    expect(alicePlaintext).toEqual(alicePlaintext);
   });
 });
