+++ packages/pre/test/message-kit.test.ts
@@ -1,7 +1,8 @@
-import { fakeBob } from '@nucypher/test-utils';
 import { expect, test } from 'vitest';
 
 import { MessageKit, toBytes } from '../src';
+
+import { fakeBob } from './utils';
 
 test('message kit', () => {
   test('bob decrypts', () => {
