+++ packages/shared/src/index.ts
@@ -1,14 +1,8 @@
-export * from './cohort';
-export * from './conditions';
 export * from './contracts';
 export * from './porter';
 export * from './types';
 export * from './utils';
 export * from './web3';
-
-import * as conditions from './conditions';
-
-export { conditions };
 
 // Re-exports
 export {
