+++ packages/taco/package.json
@@ -38,11 +38,15 @@
     "typedoc": "typedoc"
   },
   "dependencies": {
+    "@ethersproject/abstract-signer": "^5.7.0",
     "@nucypher/nucypher-core": "0.13.0-alpha.1",
-    "@nucypher/shared": "workspace:*"
+    "@nucypher/shared": "workspace:*",
+    "semver": "^7.5.2",
+    "zod": "^3.22.1"
   },
   "devDependencies": {
-    "@nucypher/test-utils": "workspace:*"
+    "@nucypher/test-utils": "workspace:*",
+    "@types/semver": "^7.5.0"
   },
   "peerDependencies": {
     "ethers": "^5.7.2"
