+++ packages/taco/src/conditions/index.ts
@@ -1,8 +1,6 @@
-// TODO: Do we want structured exports in @nucypher/shared?
 import * as base from './base';
 import * as predefined from './predefined';
 
-// TODO: Or do we want to export everything from the base and predefined modules?
 export * from './base';
 export * from './predefined';
